@@ -1,3 +1,3 @@
 module github.com/creachadair/staticfile
 
-go 1.12
+go 1.16
